refactor(mcp-server): extract token generation logic into build-tokens.ts
